requirements.txt
annotated-types==0.5.0
anyio==3.7.1
astroid==2.15.6
bcrypt==4.0.1
black==23.7.0
certifi==2023.7.22
cffi==1.15.1
click==8.1.6
cryptography==41.0.3
dill==0.3.7
dnspython==2.4.1
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.101.0
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
lint==1.2.1
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
orjson==3.9.2
packaging==23.1
passlib==1.7.4
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.2.0
protobuf==4.21.12
pyasn1==0.5.0
pycparser==2.21
pydantic==2.1.1
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydantic_core==2.4.0
pylint==2.17.5
pytest==7.4.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==1.4.23
starlette==0.27.0
tomlkit==0.12.1
typing_extensions==4.7.1
ujson==5.8.0
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.15.0
